server/requirements.txt
openai-whisper>=20231117
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
numpy>=1.24.0
requests>=2.31.0
